ci: add check commits job
* ci: add check commits job

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* ci: testing pipeline

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* ci: testing

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* ci: relates to #121 commit checks

added ability to check the latest commit on push. this
allows for developer to get quick feedback on a bad
commit message and adjust it prior to the contxt of
a PR.  additionally, within a PR, a developer will
have all commits check prior to a final merge.

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* ci: quick refactor prior to merge

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* ci: add check commits job

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* ci: testing pipeline

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* ci: testing

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* ci: relates to #121 commit checks

added ability to check the latest commit on push. this
allows for developer to get quick feedback on a bad
commit message and adjust it prior to the contxt of
a PR.  additionally, within a PR, a developer will
have all commits check prior to a final merge.

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* ci: quick refactor prior to merge

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* ci: added debugging

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* refactor: use get-pr-commits action with regex

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* ci: remove debug logging

Signed-off-by: Dustin Scott <sdustin@vmware.com>