fix: fixes #95 correct plural names (#156)
* fix: adjust test cases to new paths for vscode

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* refactor: use new upstream kustomization template

This commit pulls in the latest kustomization template from kubebuilder.
This allows us to use markers to insert resources as well reuse the
community template. We were using the community template before, but
it appears to since have been outdated.

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* fix: fixes #95, use controller tools pluralize

This commit uses the pluralize method from the kubebuilder package
rather than our own internal pluralize method.  This also would
have caused a bug with rbac as well, as the resources were not
pluralized correctly.

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* refactor: remove unneeded pluralize method

This pluralize method introduced a bug in that it would not
correctly pluralize custom resource names.

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* fix: fix missing pluralizekind method on collection resource

Signed-off-by: Dustin Scott <sdustin@vmware.com>