refactor: use external repo for resource checking (#196)
* refactor: use external resources repo

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* fix: added missing gvk before creating resource

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* chore: fix linter errors

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* fix: removed logic that was duplicated from the controller loop

This caused problems on standalone workloads that would be running
through the logic to check to see if multiple collections exist
(unsupported at this time) and error out if more than 1 exists.
Each workload would be seen as a collection and it would limit
only one workload per cluster.

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* fix: unused packages

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* fix: removed resources from reconciler

This caused issues when multiple resources were submitted.  In order to keep
the resources stored on the controller, we would have to implement a nasty
nested namespaced list of them, which would likely spike memory for clusters
which had lots of objects.  Falling back to old methodology of pulling these
resources inline.

Fixes #200
Fixes #199

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* refactor: get current state from cluster before checking ready

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* fix: removed unused packages

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* fix: remove hard-coded package name

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* fix: missing logic for component vs collection

Signed-off-by: Dustin Scott <sdustin@vmware.com>