fix: Fixes #256, added ability to associate all fields after processing (#263)
* fix: Fixes #236 glob pattern for doublestar

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* fix: fix panic when workload references collectionField

This commit addresses a panic situation when the workload itself references a
collection field, however the workload itself is not using a collection field
as an input to any of its fields.

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* test: add functional test for panic situation

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* fix: Fixes #256, added ability to associate all fields after processing

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* fix: handle array to interface conversion more intelligently

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* fix: handle case of comments below markers being seen as markers

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* test: use namespaced resource instead of cluster scoped resource

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* chore: fix linter error by removing linting item for TODO

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* chore: prepare test case for #271

This simply is a reminder to prepare for a test case for #271.  It does not
include an empty line, as an empty line would break the test case.  Once
in order to produce our functional test to account for this.

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* fix: Fixes #272, unmarshal struct with omitted fields

This adds the 'omitempty' tag to all fields within the API.  This allows the
unmarshal to unmarshal with empty fields.  Without the 'omitempty' tag, the
unmarshal event results in zero values before it can be processed by the CRD
which is there to set the default values.

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* test: remove conflict between quota and deployments without resources

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* test: fix logging test when controller is not deployed in cluster

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* test: fix missing os path in generated file

Signed-off-by: Dustin Scott <sdustin@vmware.com>

* test: adjusted resources for resource quota

Signed-off-by: Dustin Scott <sdustin@vmware.com>