chore: disable linting for strings.Title
cases.Title is set to replace strings.Title, but it does not return the same.  We consider a
'title' of a field of collection.spec.name to read collection.Spec.Name, but the new
method returns collection.Spec.name instead.  Only the first field is capitalized and the
rest are lowercased.  We will use the deprecated method until we can figure out how to
make it return appropriately.

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>