feat: added improvements to child resource mutation
This adds inline functionality to mutate resources.  Prior to this change, the mutation functionality
would loop through all resources for each child resource to perform mutation.  This resulted in a lot
of if/then statements and really ugly code.  This new change mutates the resources inline and also
passes the request with context, and the reconciler so that cluster lookups and storage of data on the
context may happen between resources.

Also Fixes #17, which now makes the header comments more legible and strips the ugly start/end tags.

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>