feat: improve mutate efficiency and logic (#48)
* feat: added improvements to child resource mutation

This adds inline functionality to mutate resources.  Prior to this change, the mutation functionality
would loop through all resources for each child resource to perform mutation.  This resulted in a lot
of if/then statements and really ugly code.  This new change mutates the resources inline and also
passes the request with context, and the reconciler so that cluster lookups and storage of data on the
context may happen between resources.

Also Fixes #17, which now makes the header comments more legible and strips the ugly start/end tags.

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* feat: move names into its own package

This will prevent import cycle conflicts when attempting to retrieve names from
packages.

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* feat: Fixes #29, create source file names with intelligence

This commit addresses an issue that creates very long file names for each manifest.  This
instead prefers the file name, and works its way back through the directory structure until a
unique name is found.  If the names cannot be deduplicated, then a count is appended to the
end.

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* feat: decrease mutate file name length

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* refactor: change from names to constants package

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* chore: fix linter errors

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* fix: backwards logic for checking out of range file names

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* chore: rename constants.go to names.go to make file more specific

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* chore: add pod disruption budget alias

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* chore: add webhook aliases

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>