fix: Fixes #49, inject resource info to update PROJECT file
This commit addresses and issue where not all information for all components are
properly injected into the PROJECT file post-generation.  To fix this, we use
the underlying kubebuilder methods to update a resource, within the scaffold
loop so that the information gets properly propogated to the PROJECT file.  This
will allow us to create webhooks at a later time.

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>