fix: remove appending of descriptions from different markers
If there are more than two instances of an identical marker with a
description, an error will be returned because the descriptions are
being appended to one another.  So identical markers throw an error
indicating they are different if there are > 2 instances of the
identical marker.  This fix removes the append.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>