fix: remove appending of descriptions from different markers (#56)
* fix: remove appending of descriptions from different markers

If there are more than two instances of an identical marker with a
description, an error will be returned because the descriptions are
being appended to one another.  So identical markers throw an error
indicating they are different if there are > 2 instances of the
identical marker.  This fix removes the append.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

* fix: correct check on descriptions array

Previously, the descriptions array was being checked if nil.  However,
the array of strings is never nil.  It's either an empty array or it has
string values in the array.  Therefore, the descriptions were never
added.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

* fix: resolve linting error

Signed-off-by: Rich Lander <lander2k2@protonmail.com>

Signed-off-by: Rich Lander <lander2k2@protonmail.com>