fix: enforce go 1.22
Because a future version of go (e.g. go 1.23) is not out yet, we are on the
latest version.  For the operator-builder binary, we are not dependent on
layered go versions (e.g. generated code go versions) so we have more flexibility
in version selection.  Removed the multiple array selection, but left it as an
array in case we want to test future versions when they release, or multiple
versions for whatever the case may be.

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>