fix: add event markers and update go version (#73)
* fix: added missing events marker

This was missed when we recently added changes to add events to deployed resources.

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* feat: switch to gener8s over legacy ocgk package

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* fix: enforce go 1.22

Because a future version of go (e.g. go 1.23) is not out yet, we are on the
latest version.  For the operator-builder binary, we are not dependent on
layered go versions (e.g. generated code go versions) so we have more flexibility
in version selection.  Removed the multiple array selection, but left it as an
array in case we want to test future versions when they release, or multiple
versions for whatever the case may be.

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* test: add ability to use different binary versions versus codebase version

This is because the go binary can be built on different versions than
the codebase itself.

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* feat: use 1.22 in go mod

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* feat: add toolchain for dealing with lower go versions

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* test: add flag to prevent binary builds when requested

This is because we were duplicating the operator-builder build in the pipeline.

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* docs: update go version in docs

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* ci: update kind and kubectl versions in pipeline

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

* ci: update goreleaser and remove deprecated fields

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>

---------

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>